+++ PassTheBomb/src/com/passthebomb/view/gui/Background.java
@@ -5,30 +5,30 @@ import com.badlogic.gdx.math.Vector2;
 
 public class Background {
 	private static Background instance = null;
-	private final int LEFT_LIM = 0;
-	private final int RIGHT_LIM;
-	private final int TOP_LIM;
-	private final int BOTTOM_LIM = 0;
+	private final float LEFT_LIM = 0;
+	private final float RIGHT_LIM;
+	private final float TOP_LIM;
+	private final float BOTTOM_LIM = 0;
 	
 	private Texture bgImg;
 	private Vector2 bgPos;
 	
-	private Background(Texture bgImg, Vector2 startPos, int[] bgSize){
+	private Background(Texture bgImg, Vector2 startPos, float[] bgSize){
 		this.bgImg = bgImg;
 		this.bgPos = startPos;
 		this.RIGHT_LIM = bgSize[0];
 		this.TOP_LIM = bgSize[1];
 	}
 	
-	public static Background createBG(Texture bgImg, Vector2 startPos, int[] bgSize){
+	public static Background createBG(Texture bgImg, Vector2 startPos, float[] bgSize){
 		if (instance == null){
 			instance = new Background(bgImg, startPos, bgSize);
 		}
 		return instance;
 	}
 	
-	public int[] getLimits(){
-		int[] output = {TOP_LIM,BOTTOM_LIM,LEFT_LIM,RIGHT_LIM};
+	public float[] getLimits(){
+		float[] output = {TOP_LIM,BOTTOM_LIM,LEFT_LIM,RIGHT_LIM};
 		return output;
 	}
 	/*public void scrollX(float val){
